AccesoLab/AccesoLab/requirements.txt
absl-py==1.0.0
altgraph==0.17.2
astunparse==1.6.3
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cvzone==1.5.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
imutils==0.5.4
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
Markdown==3.3.7
mediapipe==0.10.21
msvc-runtime==14.29.30133
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==8.4.0
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyhid-usb-relay==0.1.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyusb==1.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.3
scipy==1.7.3
setuptools==59.6.0
six==1.15.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.12.1
zipp==3.6.0

